Renamed dao, permissions and ganache files
