+++ src/tasks/start/start-backend.ts
@@ -1,17 +1,17 @@
 import chokidar from 'chokidar'
 import { BuidlerRuntimeEnvironment } from '@nomiclabs/buidler/types'
-import { createDao } from './backend/dao'
+import { createDao } from './backend/create-dao'
 import { deployImplementation } from './backend/app'
 import { createProxy, updateProxy } from './backend/proxy'
 import { resolveRepo, majorBumpRepo } from './backend/repo'
-import { setAllPermissionsOpenly } from './backend/permissions'
+import { setAllPermissionsOpenly } from './backend/set-permissions'
 import { KernelInstance, RepoInstance } from '~/typechain'
 import { logBack } from '../../ui/logger'
 import { readArapp } from '../../utils/arappUtils'
 import { AragonConfig, AragonConfigHooks } from '~/src/types'
 import { TASK_COMPILE } from '../task-names'
 import deployBases from './backend/bases/deploy-bases'
-import { startGanache } from './backend/ganache'
+import { startGanache } from './backend/start-ganache'
 import { Writable } from 'stream'
 import {
   emitEvent,
